refactor: extract Narrative service into Action classes
Decompose NarrativeService (618 lines) into 7 Action classes and a shared
NarrativeSupport service under Domain/Narrative. Actions: EditTaskNarrative,
EditDayNarrative, UndoTaskNarrative, UndoDayNarrative, RegenerateTaskNarrative,
RegenerateDayNarrative, GenerateNarrativesForReport. Update ReportController
to method injection. Add guardExportable() call in export endpoint.
