+++ backend/app/Http/Controllers/ReportController.php
@@ -4,6 +4,13 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Domain\Narrative\Actions\EditDayNarrative;
+use App\Domain\Narrative\Actions\EditTaskNarrative;
+use App\Domain\Narrative\Actions\GenerateNarrativesForReport;
+use App\Domain\Narrative\Actions\RegenerateDayNarrative;
+use App\Domain\Narrative\Actions\RegenerateTaskNarrative;
+use App\Domain\Narrative\Actions\UndoDayNarrative;
+use App\Domain\Narrative\Actions\UndoTaskNarrative;
 use App\Exceptions\NoDataException;
 use App\Http\Requests\GenerateReportRequest;
 use App\Http\Requests\UpdateNarrativeRequest;
@@ -11,7 +18,6 @@ use App\Domain\GitLab\Models\Commit;
 use App\Domain\Report\Models\Report;
 use App\Domain\Settings\Models\Setting;
 use App\Domain\Bitrix24\Models\Task;
-use App\Services\Narrative\NarrativeServiceInterface;
 use App\Services\Report\PromptExportServiceInterface;
 use App\Services\Report\ReportBuilderInterface;
 use App\Services\Report\ReportExporterInterface;
@@ -25,7 +31,6 @@ class ReportController extends Controller
     public function __construct(
         private readonly ReportBuilderInterface $builder,
         private readonly ReportExporterInterface $exporter,
-        private readonly NarrativeServiceInterface $narrativeService,
         private readonly PromptExportServiceInterface $promptExportService,
     ) {
     }
@@ -73,7 +78,7 @@ class ReportController extends Controller
      *
      * @throws NoDataException
      */
-    public function generate(GenerateReportRequest $request): JsonResponse
+    public function generate(GenerateReportRequest $request, GenerateNarrativesForReport $generateNarratives): JsonResponse
     {
         /** @var array{type: string, date_from: string, date_to: string} $validated */
         $validated = $request->validated();
@@ -94,7 +99,7 @@ class ReportController extends Controller
 
         $report = $this->builder->generate($validated['type'], $dateRange);
 
-        $this->narrativeService->generateForReport($report);
+        $generateNarratives($report);
 
         return response()->json(['data' => ['id' => $report->id]], 201);
     }
@@ -112,14 +117,14 @@ class ReportController extends Controller
     /**
      * Update day narrative.
      */
-    public function updateDay(UpdateNarrativeRequest $request, Report $report, string $date): JsonResponse
+    public function updateDay(UpdateNarrativeRequest $request, Report $report, string $date, EditDayNarrative $editDay): JsonResponse
     {
         /** @var array{narrative: string} $validated */
         $validated = $request->validated();
 
         $reportDay = $report->reportDays()->where('date', $date)->firstOrFail();
 
-        $this->narrativeService->editDayNarrative($reportDay, $validated['narrative']);
+        $editDay($reportDay, $validated['narrative']);
 
         return response()->json(['message' => 'Updated']);
     }
@@ -127,14 +132,14 @@ class ReportController extends Controller
     /**
      * Update task narrative.
      */
-    public function updateTask(UpdateNarrativeRequest $request, Report $report, int $taskId): JsonResponse
+    public function updateTask(UpdateNarrativeRequest $request, Report $report, int $taskId, EditTaskNarrative $editTask): JsonResponse
     {
         /** @var array{narrative: string} $validated */
         $validated = $request->validated();
 
         $reportTask = $report->reportTasks()->where('task_id', $taskId)->firstOrFail();
 
-        $this->narrativeService->editTaskNarrative($reportTask, $validated['narrative']);
+        $editTask($reportTask, $validated['narrative']);
 
         return response()->json(['message' => 'Updated']);
     }
@@ -142,10 +147,10 @@ class ReportController extends Controller
     /**
      * Regenerate task narrative via LLM.
      */
-    public function regenerateTask(Report $report, int $taskId): JsonResponse
+    public function regenerateTask(Report $report, int $taskId, RegenerateTaskNarrative $regenerateTask): JsonResponse
     {
         $reportTask = $report->reportTasks()->where('task_id', $taskId)->firstOrFail();
-        $updated = $this->narrativeService->regenerateTask($reportTask);
+        $updated = $regenerateTask($reportTask);
 
         return response()->json(['data' => $updated]);
     }
@@ -153,10 +158,10 @@ class ReportController extends Controller
     /**
      * Regenerate day narrative via LLM.
      */
-    public function regenerateDay(Report $report, string $date): JsonResponse
+    public function regenerateDay(Report $report, string $date, RegenerateDayNarrative $regenerateDay): JsonResponse
     {
         $reportDay = $report->reportDays()->where('date', $date)->firstOrFail();
-        $updated = $this->narrativeService->regenerateDay($reportDay);
+        $updated = $regenerateDay($reportDay);
 
         return response()->json(['data' => $updated]);
     }
@@ -164,10 +169,10 @@ class ReportController extends Controller
     /**
      * Undo last task narrative change.
      */
-    public function undoTask(Report $report, int $taskId): JsonResponse
+    public function undoTask(Report $report, int $taskId, UndoTaskNarrative $undoTask): JsonResponse
     {
         $reportTask = $report->reportTasks()->where('task_id', $taskId)->firstOrFail();
-        $result = $this->narrativeService->undoTaskNarrative($reportTask);
+        $result = $undoTask($reportTask);
 
         if ($result === null) {
             return response()->json(['message' => 'No history available'], 404);
@@ -179,10 +184,10 @@ class ReportController extends Controller
     /**
      * Undo last day narrative change.
      */
-    public function undoDay(Report $report, string $date): JsonResponse
+    public function undoDay(Report $report, string $date, UndoDayNarrative $undoDay): JsonResponse
     {
         $reportDay = $report->reportDays()->where('date', $date)->firstOrFail();
-        $result = $this->narrativeService->undoDayNarrative($reportDay);
+        $result = $undoDay($reportDay);
 
         if ($result === null) {
             return response()->json(['message' => 'No history available'], 404);
@@ -215,6 +220,8 @@ class ReportController extends Controller
      */
     public function export(Report $report): BinaryFileResponse
     {
+        $report->guardExportable();
+
         $preview = $this->builder->getPreview($report);
 
         $settings = Setting::first();
