+++ backend/bootstrap/providers.php
@@ -4,7 +4,6 @@ use App\Providers\AppServiceProvider;
 use App\Providers\Bitrix24ServiceProvider;
 use App\Providers\GitLabServiceProvider;
 use App\Providers\LlmServiceProvider;
-use App\Providers\NarrativeServiceProvider;
 use App\Providers\ReportServiceProvider;
 
 return [
@@ -12,6 +11,5 @@ return [
     Bitrix24ServiceProvider::class,
     GitLabServiceProvider::class,
     LlmServiceProvider::class,
-    NarrativeServiceProvider::class,
     ReportServiceProvider::class,
 ];
